refactor: creates a new template file for Landing Page
